LPS-113934 Move util to where it is used.
